+++ src/LaravelStarter.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Ayctor\LaravelStarter;
-
-class LaravelStarter
-{
-    // Build wonderful things
-}