+++ src/Presets/Api/Graphql.php
@@ -17,19 +17,19 @@ class Graphql extends Preset
         static::info('Starts the graphql preset installation');
 
         static::info('Run composer and artisan commands');
-        //shell_exec('composer require nuwave/lighthouse -q -n');
+        shell_exec('composer require nuwave/lighthouse -q -n');
         shell_exec('composer require mll-lab/laravel-graphql-playground -q -n');
-        //shell_exec('php artisan vendor:publish --tag=lighthouse-schema -q');
-        //shell_exec('php artisan lighthouse:ide-helper -q');
-        //shell_exec('php artisan vendor:publish --tag=lighthouse-config -q');
+        shell_exec('php artisan vendor:publish --tag=lighthouse-schema -q');
+        shell_exec('php artisan lighthouse:ide-helper -q');
+        shell_exec('php artisan vendor:publish --tag=lighthouse-config -q');
         shell_exec('php artisan vendor:publish --tag=graphql-playground-config -q');
 
-        //static::info('Change the graphql path in the configuration file');
-        //static::replaceFileValue(
-        //    config_path('lighthouse.php'),
-        //    "'uri' => '/graphql'",
-        //    "'uri' => '/api/graphql'"
-        //);
+        static::info('Change the graphql path in the configuration file');
+        static::replaceFileValue(
+            config_path('lighthouse.php'),
+            "'uri' => '/graphql'",
+            "'uri' => '/api/graphql'"
+        );
 
         static::info('Change the graphql playground path in the configuration file');
         static::replaceFileValue(