+++ stubs/base/tailwind.config.js
@@ -1,7 +1,6 @@
 module.exports = {
     theme: {
         extend: {
-            fontSize: {},
             inset: {
                 '1/4': '25%',
                 '1/2': '50%',
@@ -11,7 +10,6 @@ module.exports = {
                 '2xl': '1441px',
                 '3xl': '1920px',
             },
-            spacing: {},
             zIndex: {
                 '-1': '-1',
             }
@@ -23,28 +21,20 @@ module.exports = {
             white: '#ffffff',
             black: '#000000',
         },
-        fontFamily: {
-            'sans': [
-                'system-ui',
-                '-apple-system',
-                'BlinkMacSystemFont',
-                '"Segoe UI"',
-                '"Helvetica Neue"',
-                'Arial',
-                '"Noto Sans"',
-                'sans-serif',
-                '"Apple Color Emoji"',
-                '"Segoe UI Emoji"',
-                '"Segoe UI Symbol"',
-                '"Noto Color Emoji"',
-            ],
-        },
     },
+
     variants: {},
+
     plugins: [],
+
+    purge: [
+        './storage/framework/views/*.php',
+        './resources/views/**/*.blade.php',
+        './resources/js/**/*.vue',
+    ],
+
     future: {
         removeDeprecatedGapUtilities: true,
         purgeLayersByDefault: true,
     },
-    purge: false,
 };