+++ stubs/base/webpack.mix.js
@@ -1,57 +1,46 @@
 const mix = require('laravel-mix');
-const tailwindcss = require('tailwindcss');
 require('@ayctor/laravel-mix-svg-sprite');
 
-let options = {
-    processCssUrls: false,
-    postCss: [tailwindcss('./tailwind.config.js')],
-};
-
-mix.webpackConfig({
-    module: {
-        rules: [
-            {
-                enforce: 'pre',
-                exclude: /node_modules/,
-                loader: 'eslint-loader',
-                test: /\.(js|vue)?$/,
+mix.js('resources/js/app.js', 'public/build')
+    .postCss('resources/sass/app.scss', 'public/build', [
+        require('postcss-import'),
+        require('tailwindcss'),
+        require('postcss-nested'),
+    ])
+    .svgSprite('resources/svg/*.svg', {
+        output: {
+            filename: 'build/svg/sprite.svg',
+            chunk: {
+                name: '/build/svg/sprite',
+                keep: true,
             },
-        ],
-    },
-})
-
-if (process.NODE_ENV == 'production') {
-    options.purifyCss = true;
-}
-
-mix.js('resources/js/app.js', 'public/build');
-mix.sass('resources/sass/app.scss', 'public/build');
-
-mix.options(options);
-
-mix.svgSprite('resources/svg/*.svg', {
-    output: {
-        filename: 'build/svg/sprite.svg',
-        chunk: {
-            name: '/build/svg/sprite',
-            keep: true,
+            svg4everybody: true,
+            svgo: true,
         },
-        svg4everybody: true,
-        svgo: true,
-    },
-    sprite: {
-        prefix: false,
-        generate: {
-            title: false,
-        }
-    },
-});
-
-mix.browserSync({
-    proxy: process.env.APP_URL,
-    watch: true,
-});
-
-mix.sourceMaps(true, 'cheap-source-map', 'cheap-source-map');
-
-mix.version();
+        sprite: {
+            prefix: false,
+            generate: {
+                title: false,
+            }
+        },
+    })
+    .options({
+        processCssUrls: false,
+        purifyCss: process.NODE_ENV === 'production',
+    })
+    .webpackConfig({
+        module: {
+            rules: [
+                {
+                    enforce: 'pre',
+                    exclude: /node_modules/,
+                    loader: 'eslint-loader',
+                    test: /\.(js|vue)?$/,
+                },
+            ],
+        },
+    })
+    .browserSync({ proxy: process.env.APP_URL })
+    .sourceMaps(true, 'cheap-source-map', 'cheap-source-map')
+    .version()
+    .extract();